making the condor script easier + fixing a bug in STS

+++ bin/condor.sh
@@ -12,36 +12,65 @@ case $CMD in
 		;;
 esac
 
-expName=$2
-if [ -z "$expName" ]; then
-   echo "Experiment name can not be empty"
+expDir=$2
+if [ -z "$expDir" ]; then
+   echo "Experiment directory can not be empty"
    exit 1
 fi
 
-outputDir="condor/$expName/"
+outputDir="$expDir.exp/"
 
-step=$3
-if [ -z "$step" ] || [ "$step" -le "0" ]; then
-   echo "step can not be zero or empty"
-   exit 1
-fi
+#step=$3
+#if [ -z "$step" ] || [ "$step" -le "0" ]; then
+#   echo "step can not be zero or empty"
+#   exit 1
+#fi
 
-
-total=$(( 5000+$step-1 ))
-total=$(( $total / $step ))
-echo "Number of jobs $total"
-shift
-shift
-shift
+#total=$(( 5000+$step-1 ))
+#total=$(( $total / $step ))
+#echo "Number of jobs $total"
+#shift
+#shift
+#shift
 
 case $CMD in
-   submit)
+   submit | print)
+	  	step=$3
+	  	if [ -z "$step" ] || [ "$step" -le "0" ]; then
+		  echo "step can not be zero or empty"
+		  exit 1
+	  	fi
+
+	  	total=$(( 5000+$step-1 ))
+	  	total=$(( $total / $step ))
+	  	echo "Number of jobs $total"
+	  	shift
+		shift
+		shift
+		args="$@"
 		echo "Storing condor output files in : $outputDir. Old content will be overwritten"
-		rm $outputDir -r
-		mkdir $outputDir
-      echo "$outputDir $step $@" >> $outputDir/config
+		
+		if [ "$CMD" == "submit" ]; then
+			rm $outputDir -r
+			mkdir $outputDir
+			echo  $step  > $outputDir/config
+      	echo  $total >> $outputDir/config
+	      echo "$args" >> $outputDir/config
+		fi
+		;;
+	fix | collect)
+		step=$(head -n 1 $outputDir/config)
+	   total=$(head -n 2 $outputDir/config| tail -n 1)
+	   args=$(tail  -n 1 $outputDir/config)
+		echo "step: $step"
+		echo "total: $total"
+		echo "args: $args"
+		;;
+	*)
+		echo $CMD
 		;;
 esac
+
 
 collectBuffer=""
 for (( i=1; i<=$total; i++ ))
@@ -74,8 +103,8 @@ outputFile=$outputDir$i
 		;;
 	fix)
 		if [ "$b" == "ERROR" ]; then
-			echo "RESUBMIT bin/mlnsem run-condor $step $i $@ $outputFile"
-         bin/mlnsem run-condor $step $i $@ $outputFile
+			echo "RESUBMIT bin/mlnsem run-condor $step $i $args $outputFile"
+         bin/mlnsem run-condor $step $i $args $outputFile
 		else
 		  	echo "$i is ok"
 		fi
